% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcExcise.R
\name{calcExcise}
\alias{calcExcise}
\title{Calculates the amount of alcohol excise}
\usage{
calcExcise(abv, volume_per_bottle, bottles_per_slab = 24)
}
\arguments{
\item{abv}{Numeric. Provide the percentage alcohol by volume of the beer.}

\item{volume_per_bottle}{Numeric. Provide the volume of beer in a bottle in
mls.}

\item{bottles_per_slab}{Numeric. Provide the number of bottles of beer in a
slab. Defaults to 24.}
}
\value{
Numeric. The amount (rounded up to the nearest cent) paid in excise
    duties.
}
\description{
This function calculates the amount of excise that is payable in a given
volume of beer. Rates for 2018 are given
\href{https://www.ato.gov.au/Business/Excise-and-excise-equivalent-goods/Alcohol-excise/Excise-rates-for-alcohol/}{here}.
Note that according to the regulations, excise duty on beer is payable on the
alcohol content above 1.15% by volume.
}
