% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGST.R
\name{calcGST}
\alias{calcGST}
\title{Calculate the GST of a product.}
\usage{
calcGST(cost, gst = 0.1)
}
\arguments{
\item{cost}{Numeric. Represents the purchase price in dollars.}

\item{gst}{Numeric. Represents the percentage of GST that was added to the
purchase price. Defaults to 10% (Australia).}
}
\value{
Numeric. Returns the GST dollar amount included in the purchase
    price, rounded up to the nearest cent.
}
\description{
This simple function calculates the Goods and Services Tax (GST) component in
a given purchase price.
}
