% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAusBeerTax.R
\name{runAusBeerTax}
\alias{runAusBeerTax}
\title{Runs the Australian Beer Tax application}
\usage{
runAusBeerTax()
}
\description{
This function runs the shiny application to demonstrate the tax on your
beer.
}
